+++ VERSION-HISTORY.md
@@ -11,6 +11,7 @@
 - Add CRUD support settings and move shared code generation settings to general code generation settings section
 - Model classes from routines generated from recor result gets new name "RoutineNameResult"
 - User defined model results from a rotines (table result) are always in separate model file
+- UseStatementBody replaced UseExpressionBody
 
 ## 3.2.0